FIX: do not lose/kill sticky windows when changing layouts.
Previous behaviour caused sticky windows to get killed when deleting layouts before applying a layout and then reintroducing windows.  i3 supports sticky windows (they're great).  With this change, when we iterate through layouts (to kill) we check if a the selected window is _NET_WM_STATE_STICKY.  If so, then will not kill.